Fix implicit size_t -> int casts.

+++ src/muParserTokenReader.cpp
@@ -371,7 +371,7 @@ namespace mu
 		if (a_iPos != iEnd)
 			a_sTok = string_type(m_strFormula.begin() + a_iPos, m_strFormula.begin() + iEnd);
 
-		return iEnd;
+		return static_cast<int>(iEnd); // cast from size_t
 	}
 
 
@@ -393,7 +393,7 @@ namespace mu
 		if (a_iPos != iEnd)
 		{
 			a_sTok = string_type(m_strFormula.begin() + a_iPos, m_strFormula.begin() + iEnd);
-			return iEnd;
+			return static_cast<int>(iEnd); // cast from size_t
 		}
 		else
 		{
